Improved init command. Page tags, tag cloud page.
Tags that appear on blog pages now lead to the resp. fragment
on the page specified in the variable tagPage in the website
config. Enable showTagCloud on that page to generate a complete
list of pages tagged with each tag with id attibutes corresponding
to that tag.

This may later be simplified with default settings for tagPage
(though that would come at the cost of being less explicit).