Replaced textutil::expand with custom templating.
The new templating engine is a rewrite of Kanryu KATO's
tmpl_parser. The main feature it introduces is plain HTML being
subject to Tcl control structures (e.g., loops).

It is simpler than its inspiration thanks to the
use of regexp and no manual string escaping using for
code generation.

While it breaks backwards compatibility old templates can be
converted to the new format in also every case by simply replacing
each occurance of '<%' with '<%!' and each 'return' with an
'lindex'.