Only cache results of explicitly marked commands.
Automatic caching of the results of all commands called via
<%! ... %> in templates led to cache not being invalidated
in at least one case that is not trivial to refactor:
when the values of the commands called depend on page
variables in ways beyond "show this page element or show
nothing".

Rather than introduce more complex rules for cache invalidation
(e.g., which variables changing should invalidate what cached
values) going on we'll mark the calls to the template procs whose
results don't vary for the files in the same directory, which also
happen to be the ones that use the most resources (the sidebar
links and the tag cloud).

Also: added a new page variable "author", updated code comments,
the README and the screenshot.