Replaced ad hoc internal data structures w/ SQLite.
This is the first step of a refactoring/limited-scope rewrite
that should improve performance and simplify operations such as
adding pages that contain posts with a certain tag. The code is
still unformatted and old comments may be out of date. There are
regressions to be fixed in this first version.

Note that this is a breaking change. Some user templates will
stop working until their procs are updated to match those
in default-procs.tcl in the project skeleton.