Better fix for collection bug. Fixed tag RSS links
Remove deleted pages from articlesToAppend. Not doing this caused
the build to fail if you forgot to set hideFromCollections on your
tag page or added tags on the blog index page. The prevous fix
was less general; this should work if a new type of collections is
added.

Fixed incorrect links to RSS feeds from tag pages.